Updated with a better file size calculator

+++ cougarbackup.h
@@ -34,7 +34,7 @@ private slots:
     void number_textfield();
     void calculate_destination();
     void get_config();
-    qlonglong compute_source_size(const QString &path);
+    double compute_source_size(const QString &path);
     void update_backup_stats();
     void get_directories();
 private:
@@ -46,7 +46,7 @@ private:
     QString backup_stats;
     QString os_str;
     QString default_directory;
-    qlonglong source_size;
+    double source_size;
     queue<QDir> folders;
 };
